requirements.txt
Flask~=3.1.0
flask_sqlalchemy
Werkzeug